# Fix lifecycle phase output