## Summary
<!-- Provide a high-level summary of the change. -->

This PR addresses the side-effects of change to using `ContainerAttach` in that the `PrefixWriter` may now be called with partial line data. The solution was to buffer container log output for lifecycle phases.

## Output
<!-- If applicable, please provide examples of the output changes. -->

> NOTE: I added the use of gnomon to ensure performance was not notably hindered.

#### Before

```
./out/pack build my-ruby-app -p ../samples/apps/java-maven/ -B cnbs/sample-builder:bionic --pull-policy never | gnomon
   1.0942s   ===> DETECTING
   0.1733s   [detector] samples/java-maven 0.0.1
   0.3198s   ===> ANALYZING
   0.0001s   [analyzer] Restoring metadata for "samples/java-maven:jdk" from app image
   0.1757s   [analyzer] Restoring metadata for "samples/java-maven:maven_m2" from cache
   0.3041s   ===> RESTORING
   0.0003s   [restorer] Restoring data for "samples/java-maven:jdk" from cache
   0.6680s   [restorer] Restoring data for "samples/java-maven:maven_m2" from cache
   0.3181s   ===> BUILDING
   0.0001s   [builder] ---> Java buildpack
   0.0193s   [builder] ---> Installing JDK
   1.6992s   [builder] ---> Running Maven Wrapper
   0.0001s   [builder] [INFO] Scanning for pro
   0.3478s   [builder] jects...
   0.0001s   [builder] [INFO]
   0.0000s   [builder] [INF
   0.0000s   [builder] O] ---
   0.0002s   [builder] ----
   0.0001s   [builder] ---
   0.0001s   [builder] -
   0.0003s   [builder] -
   0.0000s   [builder] ----
   0.0003s   [builder] ----<
   0.0001s   [builder]  io.buildpacks
   0.0001s   [builder] .exampl
   0.0001s   [builder] e:sa
   0.0001s   [builder] m
   0.0001s   [builder] ple >--
   0.0003s   [builder] -
   0.0001s   [builder] ---
   0.0002s   [builder] -
   0.0003s   [builder] -------
   0.0003s   [builder] ------
   0.0001s   [builder] [INF
   0.0002s   [builder] O] Bu
   0.0003s   [builder] ilding
   0.0001s   [builder]  samp
   0.0003s   [builder] le 0.0.1
   0.0004s   [builder] -S
   0.0004s   [builder] NAPSHOT
   0.0000s   [builder] [INFO]
   0.0002s   [builder]
   0.0001s   [builder] ----
   0.0003s   [builder] -
   0.0000s   [builder] -----
   0.0003s   [builder] --
   0.0004s   [builder] -----
   0.0001s   [builder] ---------
   0.0002s   [builder] ----
   0.0002s   [builder] --[ j
   0.0002s   [builder] ar ]-----
   0.0003s   [builder] -------
   0.0002s   [builder] -----
   0.0000s   [builder] -
   0.0004s   [builder] ---
   0.0002s   [builder] ----------
   0.4894s   [builder] --
   0.0002s   [builder] [INFO
   0.0002s   [builder] ]
   0.0010s   [builder]
   0.0000s   [builder] [IN
   0.0002s   [builder] FO
   0.0003s   [builder] ]
   0.0005s   [builder] --
   0.0000s   [builder] - ma
   0.0001s   [builder] v
   0.0004s   [builder] en
   0.0005s   [builder] -cle
   0.0000s   [builder] a
   0.0001s   [builder] n-
   0.0003s   [builder] pl
   0.0001s   [builder] ug
   0.0000s   [builder] in:
   0.0001s   [builder] 3
   0.0000s   [builder] .1
   0.0002s   [builder] .
   0.0001s   [builder] 0
   0.0001s   [builder] :c
   0.0003s   [builder] l
   0.0001s   [builder] e
   0.0004s   [builder] a
   0.0003s   [builder] n (
   0.0002s   [builder] defa
   0.0004s   [builder] u
   0.0000s   [builder] lt-
   0.0002s   [builder] c
   0.0004s   [builder] l
   0.0001s   [builder] e
   0.0001s   [builder] a
   0.0002s   [builder] n
   0.0001s   [builder] )
   0.0002s   [builder]
   0.0002s   [builder] @
   0.0004s   [builder]
   0.0000s   [builder] s
   0.0002s   [builder] a
   0.0001s   [builder] m
   0.0003s   [builder] p
   0.0001s   [builder] l
   0.0002s   [builder] e
   0.0002s   [builder]
   0.0001s   [builder] -
   0.0004s   [builder] -
   0.0002s   [builder] -
   0.1407s   [builder]
   0.0001s   [builder] [
   0.0000s   [builder] I
   0.0001s   [builder] N
   0.0002s   [builder] F
   0.0002s   [builder] O
   0.0001s   [builder] ]
   0.0002s   [builder] D
   0.0004s   [builder] e
   0.0001s   [builder] leting /workspace
   0.0150s   [builder] /target
   0.0000s   [builder] [INFO]
   0.0000s   [builder] [INFO] --- maven-resourc
   0.0001s   [builder] es-plug
   0.0001s   [builder] in:3.1.0:resou
   0.0002s   [builder] rces (defa
   0.0000s   [builder] ult
   0.0001s   [builder] -resources)
   0.0002s   [builder] @ sa
   0.1329s   [builder] mple ---
   0.0000s   [builder] [INFO] Using 'UTF-8' encoding
   0.0060s   [builder] to copy filtered resources.
   0.0001s   [builder] [INFO] Copy
   0.0193s   [builder] ing 1 resource
   0.0115s   [builder] [INFO] Copying 4 resources
   0.0000s   [builder] [INFO]
   0.1546s   [builder] [INFO] --- maven-compiler-plugin:3.8.1:compile (default-compile) @ sample ---
   0.0059s   [builder] [INFO] Changes detected - recompiling the module!
   0.6582s   [builder] [INFO] Compiling 1 source file to /workspace/target/classes
   0.0000s   [builder] [INFO]
   0.0033s   [builder] [INFO] --- maven-resources-plugin:3.1.0:testResources (default-testResources) @ sample ---
   0.0000s   [builder] [INFO] Using 'UTF-8' encoding to copy filtered resources.
   0.0001s   [builder] [INFO] skip non existing resourceDirectory /workspace/src/test/resources
   0.0000s   [builder] [INFO]
   0.0000s   [builder] [INFO] --- maven-compiler-plugin:3.8.1:testCompile (default-testCo
   0.0001s   [builder] mpile) @ sample --
   0.0038s   [builder] -
   0.0000s   [builder] [INFO] No sources to compile
   0.0000s   [builder] [INFO]
   0.1269s   [builder] [INFO] --- maven-surefire-plugin:2.22.2:test (default-test) @ sample ---
   0.0000s   [builder] [INFO] Tests are skipped.
   0.0001s   [builder] [INFO]
   0.2479s   [builder] [INFO] --- maven-jar-plugin:3.1.2:jar (default-jar) @ sample ---
   0.0744s   [builder] [INFO] Building jar: /workspace/target/sample-0.0.1-SNAPSHOT.jar
   0.0000s   [builder] [INFO]
   0.6622s   [builder] [INFO] --- spring-boot-maven-plugin:2.1.17.RELEASE:repackage (repackage) @ sample ---
   0.0000s   [builder] [INFO] Replacing main artifact with repackaged archive
   0.0001s   [builder] [INFO]
   0.0000s   [builder] [INFO] --- maven-install-plugin:
   0.0563s   [builder] 2.5.2:install (default-install) @ sample ---
   0.0008s   [builder] [INFO] Installing /workspace/target/sample-0.0
   0.0305s   [builder] .1-SNAPSHOT.jar to /home/cnb/.m2/repository/io/buildpacks/example/sample/0.0.1-SNAPSHOT/sample-0.0.1-SNAPSHOT.jar
   0.0003s   [builder] [INFO] Installing /workspace/pom.xml to /home/cnb/.m2/repository/
   0.0000s   [builder] io/buildpacks/example/sample/0.0.1-SNAPSHOT/sample-0.0
   0.0000s   [builder] .1-SNAPSHOT.pom
   0.0119s   [builder]
   0.0001s   [builder] [INFO] ------------------------------------------------------------------------
   0.0001s   [builder] [INFO]
   0.0003s   [builder] BU
   0.0006s   [builder] ILD SUCCESS
   0.0005s   [builder] [INFO] ------------------------------------------------------------------------
   0.0002s   [builder] [INFO] Total time:
   0.0008s   [builder]   3.273 s
   0.0000s   [builder] [INFO] Finished at: 2020-10-02T01:09:25Z
   0.0003s   [builder] [INFO] ---------------------------------
   0.2826s   [builder] ---------------------------------------
   1.6661s   ===> EXPORTING
   0.0867s   [exporter] Reusing layer 'samples/java-maven:jdk'
   0.0095s   [exporter] Adding 1/1 app layer(s)
   0.0011s   [exporter] Reusing layer 'launcher'
   0.0005s   [exporter] Reusing layer 'config'
   0.0006s   [exporter] Reusing layer 'process-types'
   0.0000s   [exporter] Adding label 'io.buildpacks.lifecycle.metadata'
   0.0000s   [exporter] Adding label 'io.buildpacks.build.metadata'
   0.0000s   [exporter] Adding label 'io.buildpacks.project.metadata'
   1.3379s   [exporter] Setting default process type 'web'
   0.0000s   [exporter] *** Images (71835a2d9eff):
   0.0000s   [exporter]       my-ruby-app
   0.6615s   [exporter] Reusing cache layer 'samples/java-maven:jdk'
   0.5501s   [exporter] Adding cache layer 'samples/java-maven:maven_m2'
   0.0025s   Successfully built image my-ruby-app
   0.0001s

     Total   12.6024s
```

#### After

```
./out/pack build my-ruby-app -p ../samples/apps/java-maven/ -B cnbs/sample-builder:bionic --pull-policy never | gnomon
   1.2960s   ===> DETECTING
   0.1876s   [detector] samples/java-maven 0.0.1
   0.3369s   ===> ANALYZING
   0.0001s   [analyzer] Restoring metadata for "samples/java-maven:jdk" from app image
   0.1844s   [analyzer] Restoring metadata for "samples/java-maven:maven_m2" from cache
   0.2997s   ===> RESTORING
   0.0008s   [restorer] Restoring data for "samples/java-maven:jdk" from cache
   0.7239s   [restorer] Restoring data for "samples/java-maven:maven_m2" from cache
   0.2876s   ===> BUILDING
   0.0001s   [builder] ---> Java buildpack
   0.0116s   [builder] ---> Installing JDK
   1.6536s   [builder] ---> Running Maven Wrapper
   0.3318s   [builder] [INFO] Scanning for projects...
   0.0018s   [builder] [INFO]
   0.0008s   [builder] [INFO] --------------------< io.buildpacks.example:sample >--------------------
   0.0012s   [builder] [INFO] Building sample 0.0.1-SNAPSHOT
   0.4864s   [builder] [INFO] --------------------------------[ jar ]---------------------------------
   0.0013s   [builder] [INFO]
   0.1321s   [builder] [INFO] --- maven-clean-plugin:3.1.0:clean (default-clean) @ sample ---
   0.0173s   [builder] [INFO] Deleting /workspace/target
   0.0015s   [builder] [INFO]
   0.1444s   [builder] [INFO] --- maven-resources-plugin:3.1.0:resources (default-resources) @ sample ---
   0.0072s   [builder] [INFO] Using 'UTF-8' encoding to copy filtered resources.
   0.0190s   [builder] [INFO] Copying 1 resource
   0.0081s   [builder] [INFO] Copying 4 resources
   0.0040s   [builder] [INFO]
   0.1547s   [builder] [INFO] --- maven-compiler-plugin:3.8.1:compile (default-compile) @ sample ---
   0.0057s   [builder] [INFO] Changes detected - recompiling the module!
   0.6683s   [builder] [INFO] Compiling 1 source file to /workspace/target/classes
   0.0001s   [builder] [INFO]
   0.0035s   [builder] [INFO] --- maven-resources-plugin:3.1.0:testResources (default-testResources) @ sample ---
   0.0003s   [builder] [INFO] Using 'UTF-8' encoding to copy filtered resources.
   0.0005s   [builder] [INFO] skip non existing resourceDirectory /workspace/src/test/resources
   0.0004s   [builder] [INFO]
   0.0037s   [builder] [INFO] --- maven-compiler-plugin:3.8.1:testCompile (default-testCompile) @ sample ---
   0.0002s   [builder] [INFO] No sources to compile
   0.0000s   [builder] [INFO]
   0.1246s   [builder] [INFO] --- maven-surefire-plugin:2.22.2:test (default-test) @ sample ---
   0.0001s   [builder] [INFO] Tests are skipped.
   0.0004s   [builder] [INFO]
   0.2492s   [builder] [INFO] --- maven-jar-plugin:3.1.2:jar (default-jar) @ sample ---
   0.0719s   [builder] [INFO] Building jar: /workspace/target/sample-0.0.1-SNAPSHOT.jar
   0.0001s   [builder] [INFO]
   0.6320s   [builder] [INFO] --- spring-boot-maven-plugin:2.1.17.RELEASE:repackage (repackage) @ sample ---
   0.0000s   [builder] [INFO] Replacing main artifact with repackaged archive
   0.0000s   [builder] [INFO]
   0.0330s   [builder] [INFO] Installing /workspace/target/sample-0.0.1-SNAPSHOT.jar to /home/cnb/.m2/repository/io/buildpacks/example/sample/0.0.1-SNAPSHOT/sample-0.0.1-SNAPSHOT.jar
   0.0119s   [builder] [INFO] Installing /workspace/pom.xml to /home/cnb/.m2/repository/io/buildpacks/example/sample/0.0.1-SNAPSHOT/sample-0.0.1-SNAPSHOT.pomsample-0.0.1-SNAPSHOT.jar
   0.0000s   [builder] [INFO] ------------------------------------------------------------------------
   0.0006s   [builder] [INFO] BUILD SUCCESS
   0.0004s   [builder] [INFO] ------------------------------------------------------------------------
   0.0005s   [builder] [INFO] Total time:  3.217 s
   0.0003s   [builder] [INFO] Finished at: 2020-10-02T01:11:35Z
   0.2526s   [builder] [INFO] ------------------------------------------------------------------------
   1.1172s   ===> EXPORTING
   0.0731s   [exporter] Reusing layer 'samples/java-maven:jdk'
   0.0086s   [exporter] Adding 1/1 app layer(s)
   0.0005s   [exporter] Reusing layer 'launcher'
   0.0001s   [exporter] Reusing layer 'config'
   0.0005s   [exporter] Reusing layer 'process-types'
   0.0000s   [exporter] Adding label 'io.buildpacks.lifecycle.metadata'
   0.0000s   [exporter] Adding label 'io.buildpacks.build.metadata'
   0.0000s   [exporter] Adding label 'io.buildpacks.project.metadata'
   1.2169s   [exporter] Setting default process type 'web'
   0.0000s   [exporter] *** Images (6ea0b7d860d0):
   0.0006s   [exporter]       my-ruby-app
   0.5758s   [exporter] Reusing cache layer 'samples/java-maven:jdk'
   0.5784s   [exporter] Adding cache layer 'samples/java-maven:maven_m2'
   0.0019s   Successfully built image my-ruby-app
   0.0001s

     Total   11.9835s
```

## Documentation
<!-- If this change should be documented, please create an issue or PR on https://github.com/buildpacks/docs and link below. -->
<!-- NOTE: This can be added (by editing the issue) after the PR is opened. -->

- Should this change be documented?
    - No

## Related
<!-- If this PR addresses an issue, please provide issue number below. -->

Resolves #865